+++ 2.Scripts/Player.cs
@@ -6,9 +6,9 @@ using UnityEngine.UI;
 public class Player : MonoBehaviour
 {
     public int count = 0;
-    public int maxHealth;
+    public float maxHealth;
     // 최대 체력
-    public int curHealth;
+    public float curHealth;
     // 현재 체력
 
     public int Job_Index;
